+++ src/ProcMonUtils/AddressRange.cs
@@ -9,11 +9,11 @@ public interface IAddressRange
 
 public readonly struct AddressRange
 {
-    public readonly ulong Base;
-    public readonly uint Size;
-    public ulong End => Base + Size;
+    public readonly long Base;
+    public readonly int Size;
+    public long End => Base + Size;
 
-    public AddressRange(ulong @base, uint size)
+    public AddressRange(long @base, int size)
     {
         Base = @base;
         Size = size;
@@ -22,7 +22,7 @@ public readonly struct AddressRange
 
 public static class AddressRangeExtensions
 {
-    public static bool TryFindAddressIn<T>(this T[] items, ulong address, [NotNullWhen(returnValue: true)] out T? result)
+    public static bool TryFindAddressIn<T>(this T[] items, long address, [NotNullWhen(returnValue: true)] out T? result)
         where T : IAddressRange
     {
         if (items.Length == 0 || address < items[0].AddressRef.Base || address >= items[^1].AddressRef.End)
@@ -31,7 +31,7 @@ public static class AddressRangeExtensions
             return false;
         }
 
-        for (ulong l = 0, h = (ulong)(items.Length - 1); l <= h; )
+        for (long l = 0, h = items.Length - 1; l <= h; )
         {
             var i = l + (h - l) / 2;
             var test = items[i];
